chore: add changeset and rename files
