+++ .changeset/long-rabbits-camp.md
@@ -1,0 +1,6 @@
+---
+'@redocly/openapi-core': patch
+'@redocly/cli': patch
+---
+
+Updated Redocly config validation.
