+++ packages/core/src/types/redocly-yaml.ts
@@ -1,5 +1,5 @@
-import { rootRedoclyConfigSchema, apiConfigSchema } from './config-external-schemas';
-import { themeConfigSchema } from './config-external-theme-schema';
+import { rootRedoclyConfigSchema, apiConfigSchema } from './portal-config-schema';
+import { themeConfigSchema } from './theme-config';
 import { NodeType, listOf } from '.';
 import { Oas3_1Types } from './oas3_1';
 import { omitObjectProps, pickObjectProps, isCustomRuleId } from '../utils';
