+++ packages/core/src/types/theme-config.ts
